# Fix #2358, adds TIME module command to set CFE_TIME_Print() format