@CDKnightNASA 2023-06-15
Two issues raised in the CCB:
1) "%%f" would not work properly (should be replaced with "%f" in the formatted string.)
2) strstr is fairly expensive, is there a more efficient way?
  a) remove the "%f" logic and have a flag in CFE_TIME_Global to tack microseconds at the end.
  b) do a one-time scan (at startup, at command) to set offset(s) for the %f(s)