@dmknutsen 2023-09-14
I was able to perform some initial testing and think we are on track to merge into Equuleus-rc2.  I did run into a couple issues that I would like to resolve prior to merging:

1. We should make the default format backward compatible with the existing timestamp (there was a '-' removed between the Julian Day and Hour)
2. Because CFE_TIME_Print is called so often - I'd like to benchmark the new logic/while loop against a GR712 (which we are currently setting up in the cFS lab or CI/CD + bench-marking performance) to make sure there are not performance implications.
3. We will need to submit a separate PR to ensure that the example mission and platform config files are kept in sync (we added those toward the end of the Equuleus-rc1 dev cycle to demonstrate backward compatibility).
4. I'm seeing timestamp corruption prior to cFE_TIME initialization (after which they are fine).  I'd like to understand if that is an actual bug or if it is related to how I merged the PR into my test environment.