@dzbaker 2023-11-30
CCB 30 November 2023: Going to run performance tests to make sure that changes don't result in a performance hit. Afterwards, backwards compatibility issues will be fixed.