@dmknutsen 2023-12-08
@CDKnightNASA - I set up a test to benchmark this against a GR712 running RTEMS 5.  During the test, I called CFE_TIME_Print 10,000 times to see how long it takes with and without the patch.  I ran the test three times to get an average.

Without patch:
EVS Port1 1980-012-14:03:27.26382 66/1/CFE_TEST_APP 5: [  MIR] 01.005 time_misc_test.c:66 - Elapsed time for Time Print test: 1013326 usec
EVS Port1 1980-012-14:03:27.25918 66/1/CFE_TEST_APP 5: [  MIR] 01.005 time_misc_test.c:66 - Elapsed time for Time Print test: 1013291 usec
EVS Port1 1980-012-14:03:27.27995 66/1/CFE_TEST_APP 5: [  MIR] 01.005 time_misc_test.c:66 - Elapsed time for Time Print test: 1013685 usec
 
With patch:
EVS Port1 1980-012 14:03:27.53286 66/1/CFE_TEST_APP 5: [  MIR] 01.005 time_misc_test.c:66 - Elapsed time for Time Print test: 1312445 usec
EVS Port1 1980-012 14:03:27.56680 66/1/CFE_TEST_APP 5: [  MIR] 01.005 time_misc_test.c:66 - Elapsed time for Time Print test: 1313442 usec
EVS Port1 1980-012 14:03:27.54457 66/1/CFE_TEST_APP 5: [  MIR] 01.005 time_misc_test.c:66 - Elapsed time for Time Print test: 1313067 usec

Against the GR712, the time to execute CFE_TIME_Print is 28% slower with the patch - which i do not believe is cause for concern.  In this case, the enhanced functionality is most likely worth the performance trade.  In practice the performance loss will be negligible as CFE_TIME_Print should never be called at a rate where this loss would be noticeable/impactful.

Lets plan to discuss/get concurrence at the next CCB.