@dzbaker 2024-01-04
CCB 4 January 2024: Determined that formatting can be applied at compile-time.